refactor(go): split tests into standalone main server and test client
