+++ samples/go/extensions/timestamp/tests/client_test.go
@@ -1,59 +1,21 @@
-package timestamp_test
+package main
 
 import (
 	"context"
 	"fmt"
-	"iter"
 	"net"
 	"net/http"
 	"testing"
 	"time"
 
-	"samples/go/extensions/timestamp"
+	"samples/go/extensions/timestamp/timestamp"
 
 	"github.com/a2aproject/a2a-go/v2/a2a"
 	"github.com/a2aproject/a2a-go/v2/a2aclient"
 	"github.com/a2aproject/a2a-go/v2/a2aclient/agentcard"
-	"github.com/a2aproject/a2a-go/v2/a2asrv"
 )
 
 var fixedTime = time.Unix(1700000000, 0).UTC()
-
-type echoExecutor struct{}
-
-var _ a2asrv.AgentExecutor = (*echoExecutor)(nil)
-
-func (e *echoExecutor) Execute(ctx context.Context, execCtx *a2asrv.ExecutorContext) iter.Seq2[a2a.Event, error] {
-	return func(yield func(a2a.Event, error) bool) {
-		if execCtx.StoredTask == nil {
-			task := a2a.NewSubmittedTask(execCtx, execCtx.Message)
-			if !yield(task, nil) {
-				return
-			}
-		}
-
-		statusMsg := a2a.NewMessage(a2a.MessageRoleAgent, a2a.NewTextPart("working..."))
-		statusUpdate := a2a.NewStatusUpdateEvent(execCtx, a2a.TaskStateWorking, statusMsg)
-		if !yield(statusUpdate, nil) {
-			return
-		}
-
-		artEvent := a2a.NewArtifactEvent(execCtx, a2a.NewTextPart("hello!"))
-		artEvent.Artifact.Name = "result"
-		if !yield(artEvent, nil) {
-			return
-		}
-
-		completedUpdate := a2a.NewStatusUpdateEvent(execCtx, a2a.TaskStateCompleted, nil)
-		if !yield(completedUpdate, nil) {
-			return
-		}
-	}
-}
-
-func (e *echoExecutor) Cancel(ctx context.Context, execCtx *a2asrv.ExecutorContext) iter.Seq2[a2a.Event, error] {
-	return func(yield func(a2a.Event, error) bool) {}
-}
 
 func TestTimestampExtensionRoundTrip(t *testing.T) {
 	ctx := context.Background()
@@ -63,6 +25,7 @@ func TestTimestampExtensionRoundTrip(t *testing.T) {
 		return fixedTime
 	}))
 
+	// Setup local in-memory server
 	listener, err := net.Listen("tcp", "127.0.0.1:0")
 	if err != nil {
 		t.Fatalf("Failed to listen: %v", err)
@@ -72,28 +35,9 @@ func TestTimestampExtensionRoundTrip(t *testing.T) {
 	baseURL := fmt.Sprintf("http://%s", listener.Addr().String())
 	serverURL := fmt.Sprintf("%s/invoke", baseURL)
 
-	card := ext.AddToCard(&a2a.AgentCard{
-		Name:        "Echo",
-		Description: "echo agent that demonstrates the timestamp extension",
-		Version:     "1.0.0",
-		SupportedInterfaces: []*a2a.AgentInterface{
-			a2a.NewAgentInterface(serverURL, a2a.TransportProtocolJSONRPC),
-		},
-		DefaultInputModes:  []string{"text"},
-		DefaultOutputModes: []string{"text"},
-		Capabilities:       a2a.AgentCapabilities{Streaming: true},
-	})
+	handler, _ := SetupServerConfig(ext, serverURL)
 
-	serverInterceptor := timestamp.NewServerInterceptor(ext)
-	handler := a2asrv.NewHandler(
-		timestamp.WrapExecutor(&echoExecutor{}, ext),
-		a2asrv.WithExecutorContextInterceptor(serverInterceptor),
-	)
-	mux := http.NewServeMux()
-	mux.Handle("/invoke", a2asrv.NewJSONRPCHandler(handler))
-	mux.Handle(a2asrv.WellKnownAgentCardPath, a2asrv.NewStaticAgentCardHandler(card))
-
-	server := &http.Server{Handler: mux}
+	server := &http.Server{Handler: handler}
 	go func() {
 		_ = server.Serve(listener)
 	}()
