+++ samples/go/extensions/timestamp/timestamp/client.go
