+++ samples/go/extensions/timestamp/timestamp/core.go
