+++ samples/go/extensions/timestamp/timestamp/server.go
